style webpage components without type selectors
gave classes and ids as required, nice. however, to avoid unnecessary
repetition of lines of code, the solutions proposed something i didn't
do (i did it for the first item, not the fourth item!) which is shared
classes! for the fourth item, i added an id and declared all the styles
under it, when i could have given it one of the classes i gave to
the 3rd element allowing me to cut down a line of code while getting
the same outcome. it's okay... although, i did edit my code to adapt the
same approach!!!